IBM-Swift/Kitura#435 Changed macOS code to use DispatchSource instead of DispatchIO
